Fix u16 overflow in cell index format and rewrite intersects query.
The cell index serialized shape count and edge count per shape as u16.
A geometry with more than 65,535 edges in a single cell silently
truncated the count, corrupting all subsequent fields in the cell and
producing garbage geometry_ids that exceeded the edge reader's
geometry_count. Widened both fields to u32 in the writer, reader, and
merger. This is a format-breaking change; existing indexes must be
rebuilt.

Widened edge_id from u16 to u32 in the clipper, interleaver, and
clipped shape. Promoted all debug_assert calls in the spatial module
to assert so the checks that would have caught these overflows run in
release mode. Added a geometry accounting assertion in the merger to
verify next_new_id matches edge_writer.geometry_count after the
interleave loop.

Rewrote the intersects query as a single-pass algorithm using bitsets
instead of a HashMap of candidates with a separate verification pass.
Each geometry is tested for containment on first encounter (both
directions: does the candidate contain the query point, does the query
contain the candidate's first vertex), then tested for edge crossings
in every cell where it appears. The candidate containment test uses a
single cached cell lookup rather than calling find per geometry.

Removed BOLO debug logging from the interleaver. Cleaned up
interleaver edge clipping to use clip_to_padded_face and
clip_edge_bound consistently with the clipper.

+++ src/spatial/clipped_shape.rs
@@ -19,7 +19,7 @@ pub struct ClippedShape {
     /// True if the geometry contains the center of the cell.
     pub contains_center: bool,
     /// Indicies of the edges stored per geometry id in the edge index.
-    pub edge_indices: Vec<u16>,
+    pub edge_indices: Vec<u32>,
 }
 
 impl ClippedShape {
@@ -33,7 +33,7 @@ impl ClippedShape {
     }
 
     /// Adds an edge index to this clipped shape.
-    pub fn add_edge(&mut self, edge_id: u16) {
+    pub fn add_edge(&mut self, edge_id: u32) {
         self.edge_indices.push(edge_id);
     }
